Rename and add missing Poll Events
