+++ Communication/Packets/Incoming/Poll/PollAnswerEvent.cs
@@ -1,0 +1,8 @@
+﻿using Plus.HabboHotel.GameClients;
+
+namespace Plus.Communication.Packets.Incoming.Poll;
+
+internal class PollAnswerEvent
+{
+    public Task Parse(GameClient session) => throw new NotImplementedException();
+}
