+++ Communication/Packets/Incoming/Poll/PollRejectEvent.cs
@@ -1,0 +1,8 @@
+﻿using Plus.HabboHotel.GameClients;
+
+namespace Plus.Communication.Packets.Incoming.Poll;
+
+internal class PollRejectEvent
+{
+    public Task Parse(GameClient session) => throw new NotImplementedException();
+}
